requirements.txt
requests
pandas
streamlit
python-dotenv
folium
streamlit-folium
